refs #22031 moved FDA error check to load_utils and reloads data
